requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
easyocr==1.7.1
filelock==3.12.3
Flask==2.3.3
gunicorn==21.2.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy_loader==0.3
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.7.10
torch==2.0.1
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
